LayoutTests/imported/w3c/web-platform-tests/css/css-sizing/stretch/stretch-max-inline-size-001-expected.txt









FAIL [data-expected-width] 1 assert_equals:
<div class="test" data-expected-width="45"></div>
width expected 45 but got 65
FAIL [data-expected-width] 2 assert_equals:
<canvas class="test" data-expected-width="45"></canvas>
width expected 45 but got 65
FAIL [data-expected-width] 3 assert_equals:
<input class="test" data-expected-width="45">
width expected 45 but got 65
FAIL [data-expected-width] 4 assert_equals:
<textarea class="test" data-expected-width="45"></textarea>
width expected 45 but got 65
FAIL [data-expected-width] 5 assert_equals:
<button class="test" data-expected-width="45"></button>
width expected 45 but got 55
FAIL [data-expected-width] 6 assert_equals:
<div class="test" style="float: left" data-expected-width="45"></div>
width expected 45 but got 65
FAIL [data-expected-width] 7 assert_equals:
<div class="test" style="position: absolute; inset-inline: 0;" data-expected-width="55"></div>
width expected 55 but got 65
FAIL [data-expected-width] 8 assert_equals:
<div class="test" style="position: absolute" data-expected-width="50"></div>
width expected 50 but got 65
FAIL [data-expected-width] 9 assert_equals:
<div class="test" style="position: absolute; inset-inline-start: 10px" data-expected-width="45"></div>
width expected 45 but got 65
FAIL [data-expected-width] 10 assert_equals:
<div class="test" style="position: absolute; inset-inline-end: 10px" data-expected-width="45"></div>
width expected 45 but got 65
FAIL [data-expected-width] 11 assert_equals:
<div class="test" style="position: absolute; inset-inline-end: 55px" data-expected-width="10"></div>
width expected 10 but got 65
FAIL [data-expected-width] 12 assert_equals:
<div class="test" style="position: absolute; inset-inline-start: 55px" data-expected-width="10"></div>
width expected 10 but got 65
FAIL [data-expected-width] 13 assert_equals:
<div class="test" style="position: absolute; inset-inline: 0;" data-expected-width="55"></div>
width expected 55 but got 65
FAIL [data-expected-width] 14 assert_equals:
<div class="test" style="position: absolute" data-expected-width="55"></div>
width expected 55 but got 65
FAIL [data-expected-width] 15 assert_equals:
<div class="test" style="position: absolute; inset-inline-start: 10px" data-expected-width="45"></div>
width expected 45 but got 65
FAIL [data-expected-width] 16 assert_equals:
<div class="test" style="position: absolute; inset-inline-end: 10px" data-expected-width="45"></div>
width expected 45 but got 65
FAIL [data-expected-width] 17 assert_equals:
<div class="test" style="position: absolute; inset-inline-end: 55px" data-expected-width="10"></div>
width expected 10 but got 65
FAIL [data-expected-width] 18 assert_equals:
<div class="test" style="position: absolute; inset-inline-start: 55px" data-expected-width="10"></div>
width expected 10 but got 65
PASS [data-expected-width] 19
FAIL [data-expected-width] 20 assert_equals:
<div class="test" data-expected-width="45"></div>
width expected 45 but got 65
FAIL [data-expected-width] 21 assert_equals:
<div class="test" data-expected-width="55"></div>
width expected 55 but got 65
FAIL [data-expected-width] 22 assert_equals:
<div class="test" data-expected-width="30"></div>
width expected 30 but got 65
PASS [data-expected-width] 23
FAIL [data-expected-width]:not([skip-second-pass]) 24 assert_equals:
<div class="test" data-expected-width="45" style="box-sizing: border-box;"></div>
width expected 45 but got 55
FAIL [data-expected-width]:not([skip-second-pass]) 25 assert_equals:
<canvas class="test" data-expected-width="45" style="box-sizing: border-box;"></canvas>
width expected 45 but got 55
FAIL [data-expected-width]:not([skip-second-pass]) 26 assert_equals:
<input class="test" data-expected-width="45" style="box-sizing: border-box;">
width expected 45 but got 55
FAIL [data-expected-width]:not([skip-second-pass]) 27 assert_equals:
<textarea class="test" data-expected-width="45" style="box-sizing: border-box;"></textarea>
width expected 45 but got 55
FAIL [data-expected-width]:not([skip-second-pass]) 28 assert_equals:
<button class="test" data-expected-width="45" style="box-sizing: border-box;"></button>
width expected 45 but got 55
FAIL [data-expected-width]:not([skip-second-pass]) 29 assert_equals:
<div class="test" style="float: left; box-sizing: border-box;" data-expected-width="45"></div>
width expected 45 but got 55
PASS [data-expected-width]:not([skip-second-pass]) 30
FAIL [data-expected-width]:not([skip-second-pass]) 31 assert_equals:
<div class="test" style="position: absolute; box-sizing: border-box;" data-expected-width="50"></div>
width expected 50 but got 55
FAIL [data-expected-width]:not([skip-second-pass]) 32 assert_equals:
<div class="test" style="position: absolute; inset-inline-start: 10px; box-sizing: border-box;" data-expected-width="45"></div>
width expected 45 but got 55
FAIL [data-expected-width]:not([skip-second-pass]) 33 assert_equals:
<div class="test" style="position: absolute; inset-inline-end: 10px; box-sizing: border-box;" data-expected-width="45"></div>
width expected 45 but got 55
FAIL [data-expected-width]:not([skip-second-pass]) 34 assert_equals:
<div class="test" style="position: absolute; inset-inline-end: 55px; box-sizing: border-box;" data-expected-width="10"></div>
width expected 10 but got 55
FAIL [data-expected-width]:not([skip-second-pass]) 35 assert_equals:
<div class="test" style="position: absolute; inset-inline-start: 55px; box-sizing: border-box;" data-expected-width="10"></div>
width expected 10 but got 55
PASS [data-expected-width]:not([skip-second-pass]) 36
PASS [data-expected-width]:not([skip-second-pass]) 37
FAIL [data-expected-width]:not([skip-second-pass]) 38 assert_equals:
<div class="test" style="position: absolute; inset-inline-start: 10px; box-sizing: border-box;" data-expected-width="45"></div>
width expected 45 but got 55
FAIL [data-expected-width]:not([skip-second-pass]) 39 assert_equals:
<div class="test" style="position: absolute; inset-inline-end: 10px; box-sizing: border-box;" data-expected-width="45"></div>
width expected 45 but got 55
FAIL [data-expected-width]:not([skip-second-pass]) 40 assert_equals:
<div class="test" style="position: absolute; inset-inline-end: 55px; box-sizing: border-box;" data-expected-width="10"></div>
width expected 10 but got 55
FAIL [data-expected-width]:not([skip-second-pass]) 41 assert_equals:
<div class="test" style="position: absolute; inset-inline-start: 55px; box-sizing: border-box;" data-expected-width="10"></div>
width expected 10 but got 55
PASS [data-expected-width]:not([skip-second-pass]) 42
FAIL [data-expected-width]:not([skip-second-pass]) 43 assert_equals:
<div class="test" data-expected-width="45" style="box-sizing: border-box;"></div>
width expected 45 but got 55
PASS [data-expected-width]:not([skip-second-pass]) 44
FAIL [data-expected-width]:not([skip-second-pass]) 45 assert_equals:
<div class="test" data-expected-width="30" style="box-sizing: border-box;"></div>
width expected 30 but got 55
